Move large static methods to utils file
